// node/typings/results/product.ts
// Generated by https://quicktype.io

export interface GetProductRes {
    data: Data;
}

export interface Data {
    products: Products;
    product:  Product;
}

export interface Product {
    categoryId: string;
}

export interface Products {
    items: Item[];
}

export interface Item {
    id:   string;
    name: string;
    skus: Skus[];
}

export interface Skus {
    skuName: string;
    id:      string;
}